--- lib/stripe/errors/api_connection_error.rb
@@ -1,4 +1,0 @@
-module Stripe
-  class APIConnectionError < StripeError
-  end
-end
