--- lib/stripe/errors/api_error.rb
@@ -1,4 +1,0 @@
-module Stripe
-  class APIError < StripeError
-  end
-end
