--- lib/stripe/errors/authentication_error.rb
@@ -1,4 +1,0 @@
-module Stripe
-  class AuthenticationError < StripeError
-  end
-end
