--- lib/stripe/errors/permission_error.rb
@@ -1,4 +1,0 @@
-module Stripe
-  class PermissionError < StripeError
-  end
-end
