--- lib/stripe/errors/rate_limit_error.rb
@@ -1,4 +1,0 @@
-module Stripe
-  class RateLimitError < StripeError
-  end
-end
